+++ src/britive/users.py
@@ -265,11 +265,11 @@ class Users:
         :param user_ids: The list of user ids. Will be combined with `user_id`.
         :return: List of users with a small set of attributes.
         """
-
+        if user_ids is None:
+            user_ids = []
         if user_id and user_id not in user_ids:
             user_ids.append(user_id)
         if len(user_ids) == 0:
-            raise ValueError('at least one user id is required. either provide the `user_id`, '
-                             'list of users in `user_ids`, or both.')
+            return []
 
         return self.britive.post(f'{self.base_url}/minimized-user-details', json=user_ids)