+++ tests/test_010-users.py
@@ -160,7 +160,7 @@ def test_minimized_user_details(cached_user):
     details = britive.users.minimized_user_details(user_id=cached_user['userId'])
     assert isinstance(details, list)
     assert len(details) == 1
-    details = britive.users.minimized_user_details(user_ids=cached_user['userId'])
+    details = britive.users.minimized_user_details(user_ids=[cached_user['userId']])
     assert isinstance(details, list)
     assert len(details) == 1